Propagate long to code that parse values and calculates with them.
git-svn-id: https://pbtech-vc.med.cornell.edu/public/svn/icb/trunk/goby@26754 60cd2ccc-2b05-0410-9566-da273a940bff

+++ src/edu/cornell/med/icb/goby/alignments/DiscoverVariantIterateSortedAlignments.java
@@ -278,8 +278,8 @@ public class DiscoverVariantIterateSortedAlignments
                 //                     readIndex);
                 continue;
             }
-            int variationBases = stats.countVariationBases[readIndex - 1];
-            int referenceBases = stats.countReferenceBases[readIndex - 1];
+            long variationBases = stats.countVariationBases[readIndex - 1];
+            long referenceBases = stats.countReferenceBases[readIndex - 1];
             //  System.out.printf("readIndex=%d variationBases=%d referenceBases=%d %n",readIndexInfo.readIndex, variationBases, referenceBases);
             expectedVariationRate += variationBases;
             sum += variationBases + referenceBases;
